feat: implement favorite system to allow users to toggle tours and retrieve personal favorites

+++ controllers/favoriteController.js
@@ -1,14 +1,21 @@
 const asyncWrapper = require('../utils/asyncWrapper');
 const Favorites = require('../models/favorites');
 const httpStatus = require('../utils/httpStatusText');
+const AppError = require('../utils/appError');
+const Tour = require('../models/tours');
 
 exports.toggleFavorite = asyncWrapper(
   async (req, res, next) => {
     const { tourId } = req.params;
     const userId = req.currentUser._id;
 
+    const tourExists = await Tour.exists({ _id: tourId });
+    if (!tourExists) {
+      return next(new AppError('No tour found with that ID', 404));
+    }
+
     const existing = await Favorites.findOne({ tourId, userId });
-    
+
     if (existing) {
       await existing.deleteOne();
       return res.status(200).json({
@@ -40,7 +47,3 @@ exports.myFavorites = asyncWrapper(
     });
   }
 );
-
-
-
-
